move fragment instrumentation
