+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/instrumentation/fragment/RumFragmentLifecycleCallbacks.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.splunk.rum;
+package io.opentelemetry.rum.internal.instrumentation.fragment;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -25,16 +25,18 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import io.opentelemetry.api.trace.Tracer;
 import io.opentelemetry.rum.internal.instrumentation.activity.VisibleScreenTracker;
+import io.opentelemetry.rum.internal.instrumentation.fragment.FragmentTracer;
+
 import java.util.HashMap;
 import java.util.Map;
 
-class RumFragmentLifecycleCallbacks extends FragmentManager.FragmentLifecycleCallbacks {
+public class RumFragmentLifecycleCallbacks extends FragmentManager.FragmentLifecycleCallbacks {
     private final Map<String, FragmentTracer> tracersByFragmentClassName = new HashMap<>();
 
     private final Tracer tracer;
     private final VisibleScreenTracker visibleScreenTracker;
 
-    RumFragmentLifecycleCallbacks(Tracer tracer, VisibleScreenTracker visibleScreenTracker) {
+    public RumFragmentLifecycleCallbacks(Tracer tracer, VisibleScreenTracker visibleScreenTracker) {
         this.tracer = tracer;
         this.visibleScreenTracker = visibleScreenTracker;
     }
